LayoutTests/fast/dom/HTMLDialogElement/dialog-disabled.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.internals)
    internals.settings.setDialogElementEnabled(false);
</script>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<dialog id="mydialog"></dialog>
<script>
description("The DIALOG element should not be recognized if the dialog feature is off.");
dialog = document.getElementById('mydialog');
shouldBeTrue("typeof dialog.show == 'undefined'");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
